feat: show details of result as well, the "why"
I thought this would be complicated. I thought of somehow attaching or keeping track of who made what move.

But turns out this was simple. Who made what move is already captured in the variables. We can use it.

The "who made the move?" is answered in the conditionals we set up!